Fix race condition around DNS hijacking
We no longer have to apply without hijacking, wait for pods to become
ready, then apply with hijacking. Instead, when you first turn on
hijacking, if pods are not ready at that moment, we should put you in a
waiting for pods state. We'll move from that state to hijacking when
pods are first ready, and we'll stay there as long as hijacking is
enabled, even if pods become not ready again. Therefore, this only
affect the initial enabling of hijacking.